pyproject.toml
[tool.poetry]
name = "gridappsd-distributed-static-ybus-service"
version = "0.1.0"
description = ""
authors = ["afisher1 <andrew.fisher@pnnl.gov>"]
readme = "README.md"
packages = [{include = "dsys"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
gridappsd-field-bus = {version = "^2024.1.1a2", allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
yapf = "^0.40.2"
mypy = "^1.8.0"

[tool.yapfignore]
ignore_patterns = [
    ".venv/**",
    ".pytest_cache/**",
    "dist/**",
    "docs/**"
]

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
column_limit = 120
split_before_logical_operator = true

[tool.mypy]
show_error_context = true
pretty = true
show_column_numbers = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
